+++ packages/styles/bases/panel-switch-input.css.ts
@@ -15,8 +15,8 @@ export const panelSwitchBlockStyle = style({
   alignItems: "center",
   padding: `${THEME.spacing.xs} ${THEME.spacing.sm}`,
   fontSize: THEME.fontSize.xs,
-  borderTop: `1px solid ${THEME.color.gray["400"]}`,
-  borderBottom: `1px solid ${THEME.color.gray["400"]}`,
+  borderTop: `${THEME.borderWidth.xs} solid ${THEME.color.gray["400"]}`,
+  borderBottom: `${THEME.borderWidth.xs} solid ${THEME.color.gray["400"]}`,
   cursor: "pointer",
 });
 
@@ -25,21 +25,21 @@ export const panelSwitchBlockActiveStyle = style({
 });
 
 export const panelSwitchBlockDividerStyle = style({
-  borderRight: `1px solid ${THEME.color.gray["400"]}`,
+  borderRight: `${THEME.borderWidth.xs} solid ${THEME.color.gray["400"]}`,
 });
 
 export const panelSwitchBlockLeftStyle = style({
-  borderTop: `1px solid ${THEME.color.gray["400"]}`,
-  borderBottom: `1px solid ${THEME.color.gray["400"]}`,
-  borderLeft: `1px solid ${THEME.color.gray["400"]}`,
+  borderTop: `${THEME.borderWidth.xs} solid ${THEME.color.gray["400"]}`,
+  borderBottom: `${THEME.borderWidth.xs} solid ${THEME.color.gray["400"]}`,
+  borderLeft: `${THEME.borderWidth.xs} solid ${THEME.color.gray["400"]}`,
   borderTopLeftRadius: THEME.spacing.xs2,
   borderBottomLeftRadius: THEME.spacing.xs2,
 });
 
 export const panelSwitchBlockRightStyle = style({
-  borderTop: `1px solid ${THEME.color.gray["400"]}`,
-  borderBottom: `1px solid ${THEME.color.gray["400"]}`,
-  borderRight: `1px solid ${THEME.color.gray["400"]}`,
+  borderTop: `${THEME.borderWidth.xs} solid ${THEME.color.gray["400"]}`,
+  borderBottom: `${THEME.borderWidth.xs} solid ${THEME.color.gray["400"]}`,
+  borderRight: `${THEME.borderWidth.xs} solid ${THEME.color.gray["400"]}`,
   borderTopRightRadius: THEME.spacing.xs2,
   borderBottomRightRadius: THEME.spacing.xs2,
 });