+++ packages/wiz-ui-next/src/components/custom/form/control.vue
@@ -10,12 +10,24 @@
       :align="direction === 'horizontal' ? 'center' : undefined"
       :wrap="false"
     >
-      <WizHStack :width="labelWidth" align="center" gap="xs2" py="xs2">
+      <WizHStack
+        :width="labelWidth"
+        align="center"
+        :reverse="labelTagPosition === 'left'"
+        :justify="labelTagPosition === 'left' ? 'end' : 'start'"
+        gap="xs"
+        my="xs2"
+        :class="[
+          borderLeft && [borderLeftStyle, borderColorStyle[borderColor]],
+        ]"
+      >
         <WizText
           as="label"
           :htmlFor="htmlFor"
           :color="labelColor"
           :font-size="labelFontSize"
+          :bold="borderLeft"
+          :class="[borderLeft && borderLeftTextStyle]"
         >
           {{ label }}
         </WizText>
@@ -35,7 +47,12 @@
 </template>
 
 <script setup lang="ts">
-import { THEME } from "@wizleap-inc/wiz-ui-constants";
+import { ColorKeys, THEME } from "@wizleap-inc/wiz-ui-constants";
+import { borderColorStyle } from "@wizleap-inc/wiz-ui-styles/commons/border-color.css";
+import {
+  borderLeftStyle,
+  borderLeftTextStyle,
+} from "@wizleap-inc/wiz-ui-styles/customs/form-control.css";
 import { PropType, computed, inject, onMounted, provide, watch } from "vue";
 
 import {
@@ -72,6 +89,18 @@ const props = defineProps({
     type: String as PropType<"horizontal" | "vertical">,
     default: "horizontal",
   },
+  borderLeft: {
+    type: Boolean,
+    default: false,
+  },
+  borderColor: {
+    type: String as PropType<ColorKeys>,
+    default: "green.800",
+  },
+  labelTagPosition: {
+    type: String as PropType<"left" | "right">,
+    required: false,
+  },
 });
 
 // Form Group
@@ -79,6 +108,9 @@ const fromGroup = inject(formGroupKey);
 const labelWidth = computed(() => fromGroup?.labelWidth.value || "8rem");
 const labelColor = computed(() => fromGroup?.labelColor.value || "gray.900");
 const labelFontSize = computed(() => fromGroup?.labelFontSize.value || "md");
+const labelTagPosition = computed(
+  () => props.labelTagPosition || fromGroup?.labelTagPosition.value || "right"
+);
 
 const errorLeft = computed(() =>
   props.direction === "horizontal" ? labelWidth.value : "0"