+++ packages/wiz-ui-react/src/components/custom/form/stories/form-control.stories.tsx
@@ -75,3 +75,36 @@ export const Error: Story = {
     },
   },
 };
+
+export const BorderLeft: Story = {
+  args: {
+    label: "Label",
+    borderLeft: true,
+    children: <WizTextInput placeholder="入力してください" />,
+  },
+  parameters: {
+    docs: {
+      description: {
+        story:
+          "borderLeftを指定すると、labelの左側にボーダーが表示されます。`borderColor` で色を指定できます。default は `green.800` です。",
+      },
+    },
+  },
+};
+
+export const labelTagPosition: Story = {
+  args: {
+    label: "Label",
+    required: true,
+    labelTagPosition: "left",
+    children: <WizTextInput placeholder="入力してください" required />,
+  },
+  parameters: {
+    docs: {
+      description: {
+        story:
+          "labelTagPositionを指定すると、labelのTagの位置を変更できます。`left` または `right` を指定できます。default は `right` です。",
+      },
+    },
+  },
+};